+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## v0.1.6 (Finished porting)
+## v0.1.7 
+* [ ] New prensentation mode. Multiple revealjs presentation themes are provided.  
+* [ ] Added Wavedrom support.  
+* [ ] Added `Welcome Page`.  
+* [ ] Update `Mume` to version `0.0.7`.  
+
+## v0.1.6 
 * [ ] Separate **core library** out from this extension.
 * [x] Support Task List (TODO) for pandoc parser. Added click event for checkbox.    
 * [ ] Wavedrom support.